@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: asset_path.go
+//
+// Generated by this command:
+//
+//	mockgen -source=asset_path.go -destination=mock/asset_path.go -package=mock_cache
+//
 
 // Package mock_cache is a generated GoMock package.
 package mock_cache
@@ -10,7 +15,7 @@ import (
 	time "time"
 
 	model "github.com/abyssparanoia/rapid-go/internal/domain/model"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockAssetPath is a mock of AssetPath interface.
@@ -46,7 +51,7 @@ func (m *MockAssetPath) Get(ctx context.Context, assetKey string) (string, error
 }
 
 // Get indicates an expected call of Get.
-func (mr *MockAssetPathMockRecorder) Get(ctx, assetKey interface{}) *gomock.Call {
+func (mr *MockAssetPathMockRecorder) Get(ctx, assetKey any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Get", reflect.TypeOf((*MockAssetPath)(nil).Get), ctx, assetKey)
 }
@@ -60,7 +65,7 @@ func (m *MockAssetPath) Set(ctx context.Context, asset *model.Asset, expiration 
 }
 
 // Set indicates an expected call of Set.
-func (mr *MockAssetPathMockRecorder) Set(ctx, asset, expiration interface{}) *gomock.Call {
+func (mr *MockAssetPathMockRecorder) Set(ctx, asset, expiration any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Set", reflect.TypeOf((*MockAssetPath)(nil).Set), ctx, asset, expiration)
 }
