@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: asset.go
+//
+// Generated by this command:
+//
+//	mockgen -source=asset.go -destination=mock/asset.go -package=mock_repository
+//
 
 // Package mock_repository is a generated GoMock package.
 package mock_repository
@@ -9,7 +14,7 @@ import (
 	reflect "reflect"
 	time "time"
 
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockAsset is a mock of Asset interface.
@@ -45,7 +50,7 @@ func (m *MockAsset) GenerateReadPresignedURL(ctx context.Context, path string, e
 }
 
 // GenerateReadPresignedURL indicates an expected call of GenerateReadPresignedURL.
-func (mr *MockAssetMockRecorder) GenerateReadPresignedURL(ctx, path, expires interface{}) *gomock.Call {
+func (mr *MockAssetMockRecorder) GenerateReadPresignedURL(ctx, path, expires any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GenerateReadPresignedURL", reflect.TypeOf((*MockAsset)(nil).GenerateReadPresignedURL), ctx, path, expires)
 }
@@ -60,7 +65,7 @@ func (m *MockAsset) GenerateWritePresignedURL(ctx context.Context, contentType, 
 }
 
 // GenerateWritePresignedURL indicates an expected call of GenerateWritePresignedURL.
-func (mr *MockAssetMockRecorder) GenerateWritePresignedURL(ctx, contentType, path, expires interface{}) *gomock.Call {
+func (mr *MockAssetMockRecorder) GenerateWritePresignedURL(ctx, contentType, path, expires any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GenerateWritePresignedURL", reflect.TypeOf((*MockAsset)(nil).GenerateWritePresignedURL), ctx, contentType, path, expires)
 }
