@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: staff.go
+//
+// Generated by this command:
+//
+//	mockgen -source=staff.go -destination=mock/staff.go -package=mock_repository
+//
 
 // Package mock_repository is a generated GoMock package.
 package mock_repository
@@ -10,7 +15,7 @@ import (
 
 	model "github.com/abyssparanoia/rapid-go/internal/domain/model"
 	repository "github.com/abyssparanoia/rapid-go/internal/domain/repository"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockStaff is a mock of Staff interface.
@@ -45,7 +50,7 @@ func (m *MockStaff) Create(ctx context.Context, staff *model.Staff) error {
 }
 
 // Create indicates an expected call of Create.
-func (mr *MockStaffMockRecorder) Create(ctx, staff interface{}) *gomock.Call {
+func (mr *MockStaffMockRecorder) Create(ctx, staff any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Create", reflect.TypeOf((*MockStaff)(nil).Create), ctx, staff)
 }
@@ -60,7 +65,7 @@ func (m *MockStaff) Get(ctx context.Context, query repository.GetStaffQuery) (*m
 }
 
 // Get indicates an expected call of Get.
-func (mr *MockStaffMockRecorder) Get(ctx, query interface{}) *gomock.Call {
+func (mr *MockStaffMockRecorder) Get(ctx, query any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Get", reflect.TypeOf((*MockStaff)(nil).Get), ctx, query)
 }
