@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: staff_authentication.go
+//
+// Generated by this command:
+//
+//	mockgen -source=staff_authentication.go -destination=mock/staff_authentication.go -package=mock_repository
+//
 
 // Package mock_repository is a generated GoMock package.
 package mock_repository
@@ -10,7 +15,7 @@ import (
 
 	model "github.com/abyssparanoia/rapid-go/internal/domain/model"
 	repository "github.com/abyssparanoia/rapid-go/internal/domain/repository"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockStaffAuthentication is a mock of StaffAuthentication interface.
@@ -46,7 +51,7 @@ func (m *MockStaffAuthentication) CreateCustomToken(ctx context.Context, authUID
 }
 
 // CreateCustomToken indicates an expected call of CreateCustomToken.
-func (mr *MockStaffAuthenticationMockRecorder) CreateCustomToken(ctx, authUID interface{}) *gomock.Call {
+func (mr *MockStaffAuthenticationMockRecorder) CreateCustomToken(ctx, authUID any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateCustomToken", reflect.TypeOf((*MockStaffAuthentication)(nil).CreateCustomToken), ctx, authUID)
 }
@@ -61,7 +66,7 @@ func (m *MockStaffAuthentication) CreateIDToken(ctx context.Context, authUID, pa
 }
 
 // CreateIDToken indicates an expected call of CreateIDToken.
-func (mr *MockStaffAuthenticationMockRecorder) CreateIDToken(ctx, authUID, password interface{}) *gomock.Call {
+func (mr *MockStaffAuthenticationMockRecorder) CreateIDToken(ctx, authUID, password any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateIDToken", reflect.TypeOf((*MockStaffAuthentication)(nil).CreateIDToken), ctx, authUID, password)
 }
@@ -76,7 +81,7 @@ func (m *MockStaffAuthentication) CreateUser(ctx context.Context, param reposito
 }
 
 // CreateUser indicates an expected call of CreateUser.
-func (mr *MockStaffAuthenticationMockRecorder) CreateUser(ctx, param interface{}) *gomock.Call {
+func (mr *MockStaffAuthenticationMockRecorder) CreateUser(ctx, param any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateUser", reflect.TypeOf((*MockStaffAuthentication)(nil).CreateUser), ctx, param)
 }
@@ -91,7 +96,7 @@ func (m *MockStaffAuthentication) GetUserByEmail(ctx context.Context, email stri
 }
 
 // GetUserByEmail indicates an expected call of GetUserByEmail.
-func (mr *MockStaffAuthenticationMockRecorder) GetUserByEmail(ctx, email interface{}) *gomock.Call {
+func (mr *MockStaffAuthenticationMockRecorder) GetUserByEmail(ctx, email any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetUserByEmail", reflect.TypeOf((*MockStaffAuthentication)(nil).GetUserByEmail), ctx, email)
 }
@@ -105,7 +110,7 @@ func (m *MockStaffAuthentication) StoreClaims(ctx context.Context, authUID strin
 }
 
 // StoreClaims indicates an expected call of StoreClaims.
-func (mr *MockStaffAuthenticationMockRecorder) StoreClaims(ctx, authUID, staffClaims interface{}) *gomock.Call {
+func (mr *MockStaffAuthenticationMockRecorder) StoreClaims(ctx, authUID, staffClaims any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "StoreClaims", reflect.TypeOf((*MockStaffAuthentication)(nil).StoreClaims), ctx, authUID, staffClaims)
 }
@@ -120,7 +125,7 @@ func (m *MockStaffAuthentication) VerifyIDToken(ctx context.Context, idToken str
 }
 
 // VerifyIDToken indicates an expected call of VerifyIDToken.
-func (mr *MockStaffAuthenticationMockRecorder) VerifyIDToken(ctx, idToken interface{}) *gomock.Call {
+func (mr *MockStaffAuthenticationMockRecorder) VerifyIDToken(ctx, idToken any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "VerifyIDToken", reflect.TypeOf((*MockStaffAuthentication)(nil).VerifyIDToken), ctx, idToken)
 }
