@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: tenant.go
+//
+// Generated by this command:
+//
+//	mockgen -source=tenant.go -destination=mock/tenant.go -package=mock_repository
+//
 
 // Package mock_repository is a generated GoMock package.
 package mock_repository
@@ -10,7 +15,7 @@ import (
 
 	model "github.com/abyssparanoia/rapid-go/internal/domain/model"
 	repository "github.com/abyssparanoia/rapid-go/internal/domain/repository"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockTenant is a mock of Tenant interface.
@@ -46,7 +51,7 @@ func (m *MockTenant) Count(ctx context.Context, query repository.ListTenantsQuer
 }
 
 // Count indicates an expected call of Count.
-func (mr *MockTenantMockRecorder) Count(ctx, query interface{}) *gomock.Call {
+func (mr *MockTenantMockRecorder) Count(ctx, query any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Count", reflect.TypeOf((*MockTenant)(nil).Count), ctx, query)
 }
@@ -60,7 +65,7 @@ func (m *MockTenant) Create(ctx context.Context, tenant *model.Tenant) error {
 }
 
 // Create indicates an expected call of Create.
-func (mr *MockTenantMockRecorder) Create(ctx, tenant interface{}) *gomock.Call {
+func (mr *MockTenantMockRecorder) Create(ctx, tenant any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Create", reflect.TypeOf((*MockTenant)(nil).Create), ctx, tenant)
 }
@@ -74,7 +79,7 @@ func (m *MockTenant) Delete(ctx context.Context, id string) error {
 }
 
 // Delete indicates an expected call of Delete.
-func (mr *MockTenantMockRecorder) Delete(ctx, id interface{}) *gomock.Call {
+func (mr *MockTenantMockRecorder) Delete(ctx, id any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Delete", reflect.TypeOf((*MockTenant)(nil).Delete), ctx, id)
 }
@@ -89,7 +94,7 @@ func (m *MockTenant) Get(ctx context.Context, query repository.GetTenantQuery) (
 }
 
 // Get indicates an expected call of Get.
-func (mr *MockTenantMockRecorder) Get(ctx, query interface{}) *gomock.Call {
+func (mr *MockTenantMockRecorder) Get(ctx, query any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Get", reflect.TypeOf((*MockTenant)(nil).Get), ctx, query)
 }
@@ -104,7 +109,7 @@ func (m *MockTenant) List(ctx context.Context, query repository.ListTenantsQuery
 }
 
 // List indicates an expected call of List.
-func (mr *MockTenantMockRecorder) List(ctx, query interface{}) *gomock.Call {
+func (mr *MockTenantMockRecorder) List(ctx, query any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "List", reflect.TypeOf((*MockTenant)(nil).List), ctx, query)
 }
@@ -118,7 +123,7 @@ func (m *MockTenant) Update(ctx context.Context, tenant *model.Tenant) error {
 }
 
 // Update indicates an expected call of Update.
-func (mr *MockTenantMockRecorder) Update(ctx, tenant interface{}) *gomock.Call {
+func (mr *MockTenantMockRecorder) Update(ctx, tenant any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Update", reflect.TypeOf((*MockTenant)(nil).Update), ctx, tenant)
 }
