@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: transactable.go
+//
+// Generated by this command:
+//
+//	mockgen -source=transactable.go -destination=mock/transactable.go -package=mock_repository
+//
 
 // Package mock_repository is a generated GoMock package.
 package mock_repository
@@ -8,7 +13,7 @@ import (
 	context "context"
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockTransactable is a mock of Transactable interface.
@@ -43,7 +48,7 @@ func (m *MockTransactable) ROTx(ctx context.Context, fn func(context.Context) er
 }
 
 // ROTx indicates an expected call of ROTx.
-func (mr *MockTransactableMockRecorder) ROTx(ctx, fn interface{}) *gomock.Call {
+func (mr *MockTransactableMockRecorder) ROTx(ctx, fn any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ROTx", reflect.TypeOf((*MockTransactable)(nil).ROTx), ctx, fn)
 }
@@ -57,7 +62,7 @@ func (m *MockTransactable) RWTx(ctx context.Context, fn func(context.Context) er
 }
 
 // RWTx indicates an expected call of RWTx.
-func (mr *MockTransactableMockRecorder) RWTx(ctx, fn interface{}) *gomock.Call {
+func (mr *MockTransactableMockRecorder) RWTx(ctx, fn any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RWTx", reflect.TypeOf((*MockTransactable)(nil).RWTx), ctx, fn)
 }
