@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: asset.go
+//
+// Generated by this command:
+//
+//	mockgen -source=asset.go -destination=mock/asset.go -package=mock_service
+//
 
 // Package mock_service is a generated GoMock package.
 package mock_service
@@ -10,7 +15,7 @@ import (
 
 	model "github.com/abyssparanoia/rapid-go/internal/domain/model"
 	service "github.com/abyssparanoia/rapid-go/internal/domain/service"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockAsset is a mock of Asset interface.
@@ -45,7 +50,7 @@ func (m *MockAsset) BatchSetStaffURLs(ctx context.Context, staffs model.Staffs) 
 }
 
 // BatchSetStaffURLs indicates an expected call of BatchSetStaffURLs.
-func (mr *MockAssetMockRecorder) BatchSetStaffURLs(ctx, staffs interface{}) *gomock.Call {
+func (mr *MockAssetMockRecorder) BatchSetStaffURLs(ctx, staffs any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "BatchSetStaffURLs", reflect.TypeOf((*MockAsset)(nil).BatchSetStaffURLs), ctx, staffs)
 }
@@ -60,7 +65,7 @@ func (m *MockAsset) CreatePresignedURL(ctx context.Context, assetType model.Asse
 }
 
 // CreatePresignedURL indicates an expected call of CreatePresignedURL.
-func (mr *MockAssetMockRecorder) CreatePresignedURL(ctx, assetType, contentType interface{}) *gomock.Call {
+func (mr *MockAssetMockRecorder) CreatePresignedURL(ctx, assetType, contentType any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreatePresignedURL", reflect.TypeOf((*MockAsset)(nil).CreatePresignedURL), ctx, assetType, contentType)
 }
@@ -75,7 +80,7 @@ func (m *MockAsset) GetWithValidate(ctx context.Context, assetType model.AssetTy
 }
 
 // GetWithValidate indicates an expected call of GetWithValidate.
-func (mr *MockAssetMockRecorder) GetWithValidate(ctx, assetType, assetKey interface{}) *gomock.Call {
+func (mr *MockAssetMockRecorder) GetWithValidate(ctx, assetType, assetKey any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetWithValidate", reflect.TypeOf((*MockAsset)(nil).GetWithValidate), ctx, assetType, assetKey)
 }
