@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: staff.go
+//
+// Generated by this command:
+//
+//	mockgen -source=staff.go -destination=mock/staff.go -package=mock_service
+//
 
 // Package mock_service is a generated GoMock package.
 package mock_service
@@ -10,7 +15,7 @@ import (
 
 	model "github.com/abyssparanoia/rapid-go/internal/domain/model"
 	service "github.com/abyssparanoia/rapid-go/internal/domain/service"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockStaff is a mock of Staff interface.
@@ -46,7 +51,7 @@ func (m *MockStaff) Create(ctx context.Context, param service.StaffCreateParam) 
 }
 
 // Create indicates an expected call of Create.
-func (mr *MockStaffMockRecorder) Create(ctx, param interface{}) *gomock.Call {
+func (mr *MockStaffMockRecorder) Create(ctx, param any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Create", reflect.TypeOf((*MockStaff)(nil).Create), ctx, param)
 }
