@@ -7,7 +7,7 @@ import (
 	"github.com/volatiletech/null/v8"
 )
 
-//go:generate mockgen -source=$GOFILE -destination=mock/$GOFILE -package=mock_repository
+//go:generate go run go.uber.org/mock/mockgen -source=$GOFILE -destination=mock/$GOFILE -package=mock_repository
 type Tenant interface {
 	Get(
 		ctx context.Context,
