@@ -9,8 +9,8 @@ import (
 	"github.com/abyssparanoia/rapid-go/internal/domain/model"
 	mock_repository "github.com/abyssparanoia/rapid-go/internal/domain/repository/mock"
 	"github.com/abyssparanoia/rapid-go/internal/pkg/uuid"
-	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
+	"go.uber.org/mock/gomock"
 )
 
 func TestAssetService_CreatePresignedURL(t *testing.T) {
