@@ -6,8 +6,7 @@ import (
 	"github.com/abyssparanoia/rapid-go/internal/infrastructure/dependency"
 	"github.com/abyssparanoia/rapid-go/internal/infrastructure/environment"
 	"github.com/abyssparanoia/rapid-go/internal/pkg/logger"
-	"github.com/caarlos0/env/v9"
-	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
+	"github.com/caarlos0/env/v10"
 	"github.com/spf13/cobra"
 )
 
@@ -23,8 +22,8 @@ func NewCreateRootStaffCmd() *cobra.Command {
 				panic(err)
 			}
 
-			logger := logger.New()
-			ctx = ctxzap.ToContext(ctx, logger)
+			l := logger.New()
+			ctx = logger.ToContext(ctx, l)
 
 			d := &dependency.Dependency{}
 			d.Inject(ctx, e)
