@@ -4,8 +4,20 @@ import (
 	"context"
 
 	"github.com/blendle/zapdriver"
-	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+)
+
+type ctxMarker struct{}
+
+type ctxLogger struct {
+	logger *zap.Logger
+	fields []zapcore.Field
+}
+
+var (
+	ctxMarkerKey = &ctxMarker{}
+	nullLogger   = zap.NewNop()
 )
 
 // New ... new logger.
@@ -19,13 +31,35 @@ func New() *zap.Logger {
 
 // L ... logger.
 var L = func(ctx context.Context) *zap.Logger {
-	return ctxzap.Extract(ctx)
+	return Extract(ctx)
 }
 
-func AddFields(ctx context.Context, fields ...zap.Field) {
-	ctxzap.AddFields(ctx, fields...)
+func AddFields(ctx context.Context, fields ...zapcore.Field) {
+	l, ok := ctx.Value(ctxMarkerKey).(*ctxLogger)
+	if !ok || l == nil {
+		return
+	}
+	l.fields = append(l.fields, fields...)
+}
+
+func Extract(ctx context.Context) *zap.Logger {
+	l, ok := ctx.Value(ctxMarkerKey).(*ctxLogger)
+	if !ok || l == nil {
+		return nullLogger
+	}
+	fields := TagsToFields(ctx)
+	fields = append(fields, l.fields...)
+	return l.logger.With(fields...)
+}
+
+func TagsToFields(ctx context.Context) []zapcore.Field {
+	fields := []zapcore.Field{}
+	return fields
 }
 
 func ToContext(ctx context.Context, logger *zap.Logger) context.Context {
-	return ctxzap.ToContext(ctx, logger)
+	l := &ctxLogger{
+		logger: logger,
+	}
+	return context.WithValue(ctx, ctxMarkerKey, l)
 }
